services/ml_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
asyncpg==0.30.0
pydantic==2.11.0
httpx==0.27.2
python-multipart==0.0.20

# OCR & AI
google-cloud-vision==3.9.0
groq==0.11.0
pytesseract==0.3.13
Pillow==11.1.0
pdf2image==1.17.0
opencv-python-headless==4.10.0.84

# Data Science
pandas==2.2.3
numpy==2.2.0
scikit-learn==1.6.0
joblib==1.4.2
openpyxl==3.1.5

# Scraping
selenium==4.28.0
webdriver-manager==4.0.2
beautifulsoup4==4.13.0
lxml==5.3.0
requests==2.32.0

# Kafka
aiokafka==0.12.0

# Utils
python-dotenv==1.0.1
aiofiles==24.1.0
PyJWT==2.10.0
